requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pytest>=7.4.0
requests-mock>=1.11.0

# New dependencies for security and logging
python-dotenv>=1.0.0
tldextract>=5.1.0
flask>=3.0.0
flask-cors>=4.0.0

# Pydantic for data validation and models
pydantic>=2.5.0
pydantic[email]>=2.5.0